culinary_agent_backend/requirements.txt
smolagents
faiss-cpu
sentence-transformers
duckduckgo-search
numpy
python-dotenv
langfuse
huggingface-hub
ddgs